+++ Source/Backends/Vulkan/Layer/Source/Compiler/ShaderCompilerDebug.cpp
@@ -36,9 +36,12 @@ ShaderCompilerDebug::ShaderCompilerDebug(InstanceDispatchTable *table) : table(t
         path /= table->applicationInfo->pApplicationName;
     }
 
-    // Clear the sub-tree
-    std::error_code ignored;
-    std::filesystem::remove_all(path, ignored);
+    // Clear the sub-tree once per process
+    if (static bool once = false; !once) {
+        std::error_code ignored;
+        std::filesystem::remove_all(path, ignored);
+        once = true;
+    }
 
     // Ensure the tree exists
     CreateDirectoryTree(path);
